Rename btnHelpers for consistency
